+++ test/fakes/fake_DataHolder.h
@@ -3,7 +3,7 @@
 
 #include "DataHolder.h"
 
-void fakeSetTime(uint64_t newTime);
+void fakeSetTime(uint32_t newTime);
 void fakeSetIsThereNewData(int newState);
 void fakeSetLatestData(uint8_t newData);
 void fakeSetFileName(const char* newFileName);
